+++ src/index.js
@@ -1,3 +1,3 @@
-export { default as Interactor } from './interactor';
+export { default, default as Interactor } from './interactor';
 export { default as interactor } from './decorator';
 export * from './interactions';